refactor(debates): one challenge card, with the countdown and action on one row
Moves the countdown to the bottom left and the state and action to the bottom
right, so a card reads as who is involved, then how long is left and what to do
about it — rather than splitting the clock away from the action it informs.

The Requests tab had its own copy of this card, so rather than restyle it twice
it now uses the same component in both roles. That closes two gaps it had: a
received challenge showed only the other person's avatar, so you couldn't see
the pairing the way the sender could, and neither role animated in the way the
claim request cards beside them do.

The sender's "Waiting for a reply" header goes with it. The footer now carries
that state next to the action, and the header was repeating the "Sent" heading
directly above it.

The row wraps rather than overflows: the recipient's two buttons are wider than
the panel at its narrowest.
